Cleaned Up Roadrunner files.
